endpoint: Refactor host endpoint label initialization
[ upstream commit 7a11238496a8bf5cea5f7c7c084c90ed36057b4a ]

Refactor the initialization of the host endpoint's labels as this code
will be used in another method in a subsequent commit.
